// admin/models/api.ts
export interface CisternResponse {
  id: number;
  qrCode: string;
  city: CityResponse;
  state: StateResponse;
  type: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface CreateCisternRequest {
  qrCode: string;
  cityId: number;
  stateId: number;
  type: CisternTypeEnum;
  ownerId: number;
}

export interface UpdateCisternRequest {
  cityId: number;
  stateId: number;
  type: CisternTypeEnum;
}

export interface City extends AbstractEntity {
  name: string;
  state: State;
  isCapital: boolean;
  createdAt: Date;
  updatedAt: Date;
}

export interface CityResponse {
  id: number;
  name: string;
  isCapital: boolean;
  state: StateResponse;
  createdAt: Date;
  updatedAt: Date;
}
// ...
export interface StateResponse {
  id: number;
  name: string;
  uf: string;
  createdAt: Date;
  updatedAt: Date;
}
// ...
export interface PagedResponse<T> {
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
  data: T[];
}
// ...
export type CisternTypeEnum = "ALVENARIA" | "POLIPROPILENO" | "PEDRA";